Updating spark-gatk execution steps.

+++ scripts/run_variant_analysis_gatk_spark.sh
@@ -1,12 +1,7 @@
 #!/usr/bin/env bash
 
 if [[ $# -ne 4 ]]; then
-    echo "Usage: run_variant_analysis_gatk.sh <hs38|hs38a|hs38DH|hs37|hs37d5> <HDFS_PATH_OF_FASTQ_file1> <HDFS_PATH_OF_FASTQ_file2> <cluster size>"
-    exit
-fi
-
-if [[ ! -f "${1}.fa" ]]; then
-    echo "😡 Missing reference genome. Run setup_reference_genome.sh."
+    echo "Usage: run_variant_analysis_gatk_spark.sh <hs38|hs38a|hs38DH|hs37|hs37d5> <HDFS_PATH_OF_FASTQ_file1> <HDFS_PATH_OF_FASTQ_file2> <cluster size>"
     exit
 fi
 
@@ -19,11 +14,17 @@ INPUT_FILE=mysequence
 DATA_DIR="/mydata"
 REFERENCE=${DATA_DIR}"/"${1}".fa"
 DICT=${DATA_DIR}"/"${1}".dict"
+BWA=${DATA_DIR}"/bwa/bwa"
 GATK=${DATA_DIR}"/gatk-4.1.8.0/gatk"
 OUTPUT_PREFIX="VA-"${USER}"-result"
 
-let NUM_EXECUTORS=${3}
+let NUM_EXECUTORS=${4}
 let NUM_CORES=$(nproc)-4
+
+if [[ ! -f "${REFERENCE}" ]]; then
+    echo "😡 Missing reference genome. Run setup_reference_genome.sh."
+    exit
+fi
 
 echo "👉 Deleting files..."
 hdfs dfs -rm -r ${HDFS_PREFIX}/${INPUT_FILE}.*
@@ -40,21 +41,21 @@ ${CANNOLI_SUBMIT} --master ${SPARK_MASTER} --driver-memory ${DRIVER_MEMORY} --nu
     -executable ${BWA} -sample_id mysample -index ${REFERENCE} -sequence_dictionary ${DICT} -single -add_files
 
 echo "👉 Sorting the bam file."
-./gatk SortSamSpark -I ${HDFS_PREFIX}/${INPUT_FILE}.bam -O ${HDFS_PREFIX}/${INPUT_FILE}-sorted.bam \
-    --spark-runner SPARK  --spark-master ${SPARK_MASTER} --driver-memory ${DRIVER_MEMORY} --num-executors ${NUM_EXECUTORS} --executor-cores ${NUM_CORES} --executor-memory ${EXECUTOR_MEMORY}
+${GATK} SortSamSpark -I ${HDFS_PREFIX}/${INPUT_FILE}.bam -O ${HDFS_PREFIX}/${INPUT_FILE}-sorted.bam \
+    --spark-runner SPARK --spark-master ${SPARK_MASTER} --conf "spark.executor.cores=${NUM_CORES}" --conf "spark.executor.memory=${EXECUTOR_MEMORY}" --conf "spark.executor.instances=${NUM_EXECUTORS}"
 
 echo "👉 marking duplicates before variant calling."
-./gatk MarkDuplicatesSpark -I ${HDFS_PREFIX}/${INPUT_FILE}-sorted.bam -O ${HDFS_PREFIX}/${OUTPUT_PREFIX}-rg-sorted-final.bam \
-    --spark-runner SPARK  --spark-master ${SPARK_MASTER} --driver-memory ${DRIVER_MEMORY} --num-executors ${NUM_EXECUTORS} --executor-cores ${NUM_CORES} --executor-memory ${EXECUTOR_MEMORY}
+${GATK} MarkDuplicatesSpark -I ${HDFS_PREFIX}/${INPUT_FILE}-sorted.bam -O ${HDFS_PREFIX}/${OUTPUT_PREFIX}-rg-sorted-final.bam \
+    --spark-runner SPARK --spark-master ${SPARK_MASTER} --conf "spark.executor.cores=${NUM_CORES}" --conf "spark.executor.memory=${EXECUTOR_MEMORY}" --conf "spark.executor.instances=${NUM_EXECUTORS}" \
     --tmp-dir ${DATA_DIR}/gatk-tmp
 
 echo "👉 Running GATK HaplotypeCaller on spark for variant calling."
-${GATK_HOME}/gatk HaplotypeCallerSpark \
-    -R ${1}.fa \
+${GATK} HaplotypeCallerSpark \
+    -R ${REFERENCE} \
     -I ${HDFS_PREFIX}/${OUTPUT_PREFIX}-rg-sorted-final.bam \
-    -O ${HDFS_PREFIX}/${OUTPUT_PREFIX}-gatk-output.vcf \
-    --spark-runner SPARK  --spark-master ${SPARK_MASTER} --driver-memory ${DRIVER_MEMORY} --num-executors ${NUM_EXECUTORS} --executor-cores ${NUM_CORES} --executor-memory ${EXECUTOR_MEMORY}
+    -O ${HDFS_PREFIX}/${OUTPUT_PREFIX}-gatk-spark-output.vcf \
+    --spark-runner SPARK --spark-master ${SPARK_MASTER} --conf "spark.executor.cores=${NUM_CORES}" --conf "spark.executor.memory=${EXECUTOR_MEMORY}" --conf "spark.executor.instances=${NUM_EXECUTORS}"
 
-hdfs dfs -copyToLocal ${HDFS_PREFIX}/${OUTPUT_PREFIX}-gatk-output.vcf ${HOME}/${OUTPUT_PREFIX}-gatk-output.vcf
-echo "👉 Done with variant calling. See ${OUTPUT_PREFIX}-gatk-output.vcf file."
+hdfs dfs -copyToLocal ${HDFS_PREFIX}/${OUTPUT_PREFIX}-gatk-spark-output.vcf ${HOME}/${OUTPUT_PREFIX}-gatk-spark-output.vcf
+echo "👉 Done with variant calling. See ${OUTPUT_PREFIX}-gatk-spark-output.vcf file."
 date
